* trying to verify 64-bit validity of code, checking casts, especially when pointers are involved * first pass of converting c-style casts to c++ casts (static|const|reinterpret|dynamic) where possible * using misc.h to_string and from_string more often * removed unnecessary casts * using size_t more often, instead of casting to and from int/unsigned int/unsigned long * upped clientsmlcsharp project to warning level 4, fixed warnings

+++ SoarSuite/Core/CLI/src/cli_setstopphase.cpp
@@ -91,7 +91,7 @@ bool CommandLineInterface::DoSetStopPhase(bool setPhase, bool before, smlPhase p
 		// The "stop after" form is just a courtesy to the user in case they prefer to think that way.
 		if (!before)
 		{
-			phase = (smlPhase)(((int)phase)+1) ;
+			phase = smlPhase(phase + 1);
 			if (phase > sml_OUTPUT_PHASE)
 				phase = sml_INPUT_PHASE ;
 		}
@@ -104,7 +104,7 @@ bool CommandLineInterface::DoSetStopPhase(bool setPhase, bool before, smlPhase p
 
 	if (!before)
 	{
-		stopPhase = (smlPhase)(((int)stopPhase)-1) ;
+		stopPhase = smlPhase(stopPhase - 1);
 		if (stopPhase < sml_INPUT_PHASE)
 			stopPhase = sml_OUTPUT_PHASE ;
 	}
